리모델링한 엔티티 전부 기본 Repository 구성완료

+++ src/main/java/net/softbell/bsh/domain/repository/MemberRepo.java
@@ -7,8 +7,12 @@ import org.springframework.stereotype.Repository;
 
 import net.softbell.bsh.domain.entity.Member;
 
+/**
+ * @Author : Bell(bell@softbell.net)
+ * @Description : 회원 리포지토리 인터페이스
+ */
 @Repository
-public interface MemberRepo extends JpaRepository<Member, Long>
+public interface MemberRepo extends JpaRepository<Member, Integer>
 {
 	Optional<Member> findByUserId(String userId);
 }
